Move UI models from app module to ui module
The placement of these models was incorrect since each model is a subclass of a Qt model, a UI concept. Additionally, each model was only used in view.py in ui module
